requirements.txt
cuid2==2.0.1
Django==5.2.7
djangorestframework==3.16.1
django-filter==24.3
djangorestframework-simplejwt==5.4.0
